+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v1.3.5 - Unreleased
 ----
  * Fix type verification in resolver arguments, GraphQL Schema must do this
+ * Fix bug with redis pub/sub handler and socket timeouts
 
 v1.3.4 - 2019-05-03
 ----